requirements.txt
Django
protobuf
cryptography
requests
python-jose
